requirements.txt
kaggle>=1.6.17
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.0.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0  # Required for Selective Search in R-CNN
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
matplotlib>=3.5.0

seaborn
scikit-learn

kagglehub

jupyterlab
notebook
ipykernel
ipywidgets
jupyterlab_widgets

optuna

# R-CNN specific dependencies
pycocotools>=2.0.6  # Object detection evaluation metrics